backend/requirements.txt
fastapi
uvicorn
pydantic
bcrypt
python-jose
python-multipart
openai
tiktoken
weaviate-client
langchain
langchain-openai
langchain-weaviate
langfuse
pdfminer.six
python-docx
nltk
aiofiles
sse-starlette
tqdm
firebase-admin
aiohttp
firebase-admin